Update Application Validation

+++ Crud/Models/Objectif.cs
@@ -21,7 +21,7 @@ namespace Crud.Models
         public string Commantaire_Manager { get; set; }
         public string Commantaire_Employee { get; set; }
         public Status_Obj Status_Obj { get; set; }
-        public int Score { get; set; }
+        public decimal Score { get; set; }
         public Entretein Entretein { set; get; }
 
         public int Annee { get; set; }
